requirements.txt
#--extra-index-url https://pypi.org/simple/
#--trusted-host pypi.org

ml-trainer-sdk==0.1.5
timm
albumentations
patool
reformer-pytorch
sktime
PyWavelets
python-dotenv
aipmodel==0.2.28
boto3
huggingface_hub #to remove
clearml